requirements.txt
earthengine-api==0.1.326
google-cloud-storage
pandas
numpy
scikit-learn
python-dotenv
tqdm
geopandas
rasterio
requests
torch
streamlit
SpeechRecognition
pyaudio
sounddevice
Flask
pydub
transformers
peft
statsmodels